fix: sizing and positioning
