+++ src/app/doyleroom/page.jsx
@@ -4,7 +4,7 @@ import styles from "./components/styles.module.css";
 import { Container, Text, Box } from "@chakra-ui/react";
 import { Suspense, useEffect, useState } from "react";
 import fetchRoom from "@/pages/api/rooms/fetchRoom";
-import CldImage from "../components/CldImage";
+import CldImage from "../components/ImageComp";
 import Map from "../Map";
 
 export default function DoyleRoom() {
