+++ src/app/maanroom/page.jsx
@@ -2,7 +2,7 @@
 import styles from './components/styles.module.css'
 import Map from '../Map'
 import { Container, Text, Box } from '@chakra-ui/react'
-import CldImage from '../components/CldImage'
+import CldImage from '../components/ImageComp'
 import { Suspense, useEffect, useState } from 'react'
 import fetchRoom from '@/pages/api/rooms/fetchRoom'
