+++ src/app/seraphineroom/page.jsx
@@ -3,7 +3,7 @@ import fetchRoom from "@/pages/api/rooms/fetchRoom";
 import { Container, Box, Text } from "@chakra-ui/react";
 
 import { useEffect, useState } from "react";
-import CldImage from "../components/CldImage";
+import CldImage from "../components/ImageComp";
 import styles from "./components/styles.module.css";
 import Map from "../Map";
